Added notebook to test
